Enable global CHI watching from any namespace with WATCH_ALL_NAMESPACES Env Var
Introduce a new environment variable WATCH_ALL_NAMESPACES, enabling the Operator to monitor CHIs resources across all namespaces, regardless of the namespace in which the operator is deployed. This enhancement removes the previous limitation where the operator could only watch all namespaces when deployed in the kube-system namespace or a specified set of namespaces when configured with watch_namespaces. With this change, users now have the flexibility to deploy the Operator in any desired namespace and still retain the capability to watch CHIs cluster-wide, including those in dynamically created namespaces.

Signed-off-by: hesorchen <hesorchen@Gmail.com>